+++ lib/doc_template/template.rb
@@ -31,7 +31,7 @@ module DocTemplate
       end
     end
 
-    attr_reader :css_styles, :metadata_service, :toc
+    attr_reader :css_styles, :documents, :metadata_service, :toc
 
     class << self
       #